Pasted in code from other sources
Pasted in listhelper class from file management project, used instead of
std::list because std::list is massive overkill for this, in fact an
argument could be made for listhelper.cc being overkill

There is now parsing code as a result of this. Minimal tailoring has
been done to make the code fit the specifics of this codebase, and it is
largely unmodified.