Added better comment, return 0 where needed
The getQuizResults() function did not return zero on success like it
should have, causing unpredictable return codes and program failure
despite all of the sources and commands matching perfectly to what they
should be.

One of the comments listed a todo about figuring out whether logic was
needed in that place. The answer is "not right now" so the question has
been replaced by a description of the existance of an if statement.