Fixed formatting bugs, dupe questions, dupe option
The program would output all options for all questions under the header
for each question, instead of only showing the ones needed for that
question. This was because of an issue with how the head element of the
option list was handled. It was being copied needlessly whenever a
question would be added to the questions list, causing this problem.